Base class SelectiveTestDlgBase for tests that founded on options selection implemented.
