+++ Presentation/Testing/TranslationWordDlg.xaml
@@ -1,8 +1,9 @@
-﻿<Window x:Class="DDictionary.Presentation.Testing.TranslationWordDlg"
+﻿<base:SelectiveTestDlgBase x:Class="DDictionary.Presentation.Testing.TranslationWordDlg"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:base="clr-namespace:DDictionary.Presentation.Testing"
         mc:Ignorable="d"
         Title="Translation - word testing" Height="370" Width="600" WindowStartupLocation="CenterOwner" 
         SizeToContent="Height" ResizeMode="NoResize" ShowInTaskbar="False" WindowStyle="ToolWindow" 
@@ -97,4 +98,4 @@
             <StackPanel DockPanel.Dock="Bottom"/>
         </DockPanel>
     </Grid>
-</Window>
+</base:SelectiveTestDlgBase>
